use delegation.committer to sign commitments, not config.committer
We assume that the relay will filter SignedDelegation to only include valid ones where the proposer is scheduled for that slot in the lookahead. We also assume the URC slashes from equivocations, so the relay will only return up to a single SignedDelegation per slot. Since the gateway populates it's DB with these relay-originated SignedDelegation, we only need to ever store one per slot and can use this in the commitment_request_handler